+++ src/canisters-official-backend/src/core/api/webhooks/mod.rs
@@ -2,3 +2,4 @@ pub mod team_invites;
 pub mod directory;
 pub mod state_diffs;
 pub mod tags;
+pub mod organization;
