+++ src/canisters-official-backend/src/core/state/contacts/state.rs
@@ -3,15 +3,18 @@ pub mod state {
     use std::cell::RefCell;
     use std::collections::HashMap;
 
-    use crate::core::{state::contacts::types::Contact, types::UserID};
+    use crate::core::{state::contacts::types::Contact, types::{ICPPrincipalString, UserID}};
     
     thread_local! {
         // default is to use the api key id to lookup the api key
         pub(crate) static CONTACTS_BY_ID_HASHTABLE: RefCell<HashMap<UserID, Contact>> = RefCell::new(HashMap::new());
         // default is to use the api key id to lookup the api key
-        pub(crate) static CONTACTS_BY_ICP_PRINCIPAL_HASHTABLE: RefCell<HashMap<String, UserID>> = RefCell::new(HashMap::new());
+        pub(crate) static CONTACTS_BY_ICP_PRINCIPAL_HASHTABLE: RefCell<HashMap<ICPPrincipalString, UserID>> = RefCell::new(HashMap::new());
         // track in hashtable users list of ApiKeyIDs
         pub(crate) static CONTACTS_BY_TIME_LIST: RefCell<Vec<UserID>> = RefCell::new(Vec::new());
+        // superswap userid history
+        // HISTORY_SUPERSWAP_USERID: HashMap<OldUserID, CurrentUserID>
+        pub(crate) static HISTORY_SUPERSWAP_USERID: RefCell<HashMap<UserID, UserID>> = RefCell::new(HashMap::new());
     }
 
 }
