+++ src/canisters-official-backend/src/core/state/permissions/state.rs
@@ -48,3 +48,4 @@ pub mod state {
     }
 
 }
+
