+++ src/canisters-official-backend/src/rest/drives/route.rs
@@ -7,13 +7,6 @@ pub const DRIVES_GET_PATH: &str =                   genroute!("/drives/get/{driv
 pub const DRIVES_LIST_PATH: &str =                  genroute!("/drives/list");
 pub const DRIVES_UPSERT_PATH: &str =                genroute!("/drives/upsert");
 pub const DRIVES_DELETE_PATH: &str =                genroute!("/drives/delete");
-pub const ORG_SNAPSHOT_PATH: &str =                 genroute!("/organization/snapshot");
-pub const ORG_REPLAY_PATH: &str =                   genroute!("/organization/replay");
-pub const ORG_SEARCH_PATH: &str =                   genroute!("/organization/search");
-pub const ORG_REINDEX_PATH: &str =                  genroute!("/organization/reindex");
-pub const ORG_EXTERNAL_ID_PATH: &str =              genroute!("/organization/external_id");
-pub const ORG_TRANSFER_OWNERSHIP_PATH: &str =       genroute!("/organization/transfer_ownership");
-pub const ORG_WHOAMI_PATH: &str =                   genroute!("/organization/whoami");
 
 type HandlerEntry = (&'static str, &'static str, RouteHandler);
 
@@ -38,42 +31,6 @@ pub fn init_routes() {
             "POST",
             DRIVES_DELETE_PATH,
             |req, params| Box::pin(crate::rest::drives::handler::drives_handlers::delete_drive_handler(req, params)),
-        ),
-        (
-            "GET",
-            ORG_SNAPSHOT_PATH,
-            |req, params| Box::pin(crate::rest::drives::handler::drives_handlers::snapshot_drive_handler(req, params)),
-        ),
-        (
-            "POST",
-            ORG_REPLAY_PATH,
-            |req, params| Box::pin(crate::rest::drives::handler::drives_handlers::replay_drive_handler(req, params)),
-        ),
-        (
-            "POST",
-            ORG_SEARCH_PATH,
-            |req, params| Box::pin(crate::rest::drives::handler::drives_handlers::search_drive_handler(req, params)),
-        ),
-        (
-            "POST",
-            ORG_REINDEX_PATH,
-            |req, params| Box::pin(crate::rest::drives::handler::drives_handlers::reindex_drive_handler(req, params)),
-        ),
-        (
-            "POST",
-            ORG_EXTERNAL_ID_PATH,
-            |req, params| Box::pin(crate::rest::drives::handler::drives_handlers::external_id_drive_handler(req, params)),
-        ),
-        (
-            "POST",
-            ORG_TRANSFER_OWNERSHIP_PATH,
-            // transfering ownership requires owner call this route twice with the same body at least 24 hours apart
-            |req, params| Box::pin(crate::rest::drives::handler::drives_handlers::transfer_ownership_drive_handler(req, params)),
-        ),
-        (
-            "GET",
-            ORG_WHOAMI_PATH,
-            |req, params| Box::pin(crate::rest::drives::handler::drives_handlers::whoami_drive_handler(req, params)),
         ),
     ];
